drfLessonDetail/requirements.txt
Django>=4.2,<5
djangorestframework
psycopg2-binary
Pillow
djangorestframework-simplejwt
drf_spectacular
django-cors-headers
django-schema-graph